BUG: fix lazy arrow arithmetic + datetime kernels to match pandas
Backend review batch 4:

- floor_divide / modulo (arrow/core.py): pc.divide truncates integer division
  toward zero and pc.floor on an int is a no-op, so -7 // 3 gave -2 (not -3)
  and -7 % 3 gave -1 (not 2), plus float dtype for int inputs. Now floor the
  real (float) quotient toward -inf and restore int dtype for int//int; modulo
  reuses it so the result follows the divisor's sign. Removed the stale
  ARROW_FUNC_MAP "floor_divide"/"mod" entries (no such pyarrow functions).
- dt_is_month_end (arrow/datetime.py): compared day-of-month to days
  *remaining* in the month, so it was ~always False; now compares to
  days_in_month.
- dt_microsecond (arrow/datetime.py): pc.microsecond is the microsecond within
  the millisecond (0-999); pandas .dt.microsecond is within the second
  (0-999999) = millisecond*1000 + microsecond. Fixed dt_nanosecond docstring.
- n_unique (arrow/core.py): len(pc.unique) counted null as a value (off-by-one
  vs pandas nunique dropna=True); now count_distinct(mode="only_valid").
- median (arrow/core.py): pc.approximate_median (t-digest) replaced with exact
  np.nanmedian.
- dt_year (numpy/datetime.py): raw datetime64[Y].astype(int) produced garbage
  for NaT and assumed ns resolution; now uses the DatetimeIndex path like the
  other dt_* kernels (int32, NaT-safe), matching eager.

Adds TestArrowArithmeticAndDatetime; updates two dt_year tests to the
eager-matching int32 dtype. Full lazy suite green (1918 passed); all 22 TPC-H
SF-3 queries validate exact (geo-mean 0.49x).

+++ pandas/tests/lazy/test_array_eval_coverage.py
@@ -406,7 +406,7 @@ class TestArrayEvaluatorDatetime:
 
         node = Call("dt_year", args=(FieldRef("date"),))
         result = evaluator.evaluate(node)
-        tm.assert_numpy_array_equal(result, np.array([2020, 2021]))
+        tm.assert_numpy_array_equal(result, np.array([2020, 2021], dtype="int32"))
 
 
 class TestArrayEvaluatorCumulative:
